Remove Route class
Move verbs and flags support to Router
