+++ test/core.router.js
@@ -1,7 +1,9 @@
 /*global describe, it*/
 'use strict';
 
+var _ = require('lodash-node');
 var assert = require('assert');
+var util = require('util');
 
 describe('core/router', function () {
     /*eslint max-nested-callbacks: 0*/
@@ -50,13 +52,15 @@ describe('core/router', function () {
                 {
                     args: {},
                     data: {
-                        name: 'index0'
+                        name: 'index0',
+                        verbs: ['GET', 'HEAD']
                     }
                 },
                 {
                     args: {},
                     data: {
-                        name: 'index1'
+                        name: 'index1',
+                        verbs: ['GET', 'HEAD']
                     }
                 }
             ]);
@@ -73,4 +77,77 @@ describe('core/router', function () {
             assert.deepEqual(router.matchVerbs('/').sort(), ['HEAD', 'GET', 'POST'].sort());
         });
     });
+
+    describe('{Router}.addRule()', function () {
+
+        describe('{Route}.params', function () {
+            it('Should support flags', function () {
+                var router = new Router();
+                var route = router.addRule('/ IzXs', {
+                    ignoreCase: true,
+                    appendSlash: false
+                });
+
+                assert.ok(!route.params.ignoreCase);
+                assert.ok(route.params.appendSlash);
+                assert.ok(route.params.z);
+                assert.ok(!route.params.X);
+            });
+
+            it('Should merge router params with route.params', function () {
+                var router = new Router({
+                    ignoreCase: true
+                });
+
+                var route = router.addRule('/ s');
+                assert.ok(route.params.ignoreCase);
+                assert.ok(route.params.appendSlash);
+            });
+        });
+
+        describe('{Route}.data.verbs', function () {
+            it('Should parse verbs', function () {
+                var router = new Router();
+                var route = router.addRule('POST /');
+                assert.deepEqual(route.data.verbs.sort(), ['POST'].sort());
+            });
+
+            it('Should automatically add HEAD to GET routes', function () {
+                var router = new Router();
+                var route = router.addRule('GET /');
+                assert.deepEqual(route.data.verbs.sort(), ['GET', 'HEAD'].sort());
+            });
+
+            it('Should add GET automatically if no verbs specified', function () {
+                var router = new Router();
+                var route = router.addRule('/');
+                assert.deepEqual(route.data.verbs.sort(), ['GET', 'HEAD'].sort());
+            });
+
+            it('Should ignore dupes in verbs list', function () {
+                var router = new Router();
+                var route = router.addRule('POST,POST, PUT, POST /');
+                assert.deepEqual(route.data.verbs.sort(), ['POST', 'PUT'].sort());
+            });
+        });
+    });
+
+    Router.parseRequestRule = Router._parseRequestRule;
+
+    describe('Router._parseRequestRule', function () {
+        var title = 'Should parse %j to %j';
+        var samples = [
+            [
+                '/<page>/?page',
+                ['', '/<page>/?page', '']
+            ]
+        ];
+        _.forEach(samples, function (s) {
+            var shouldText = util.format(title, s[0], s[1]);
+            it(shouldText, function () {
+                assert.deepEqual(Router.parseRequestRule(s[0]), s[1]);
+            });
+        });
+    });
+
 });
